Added support of multicurrency paymnets

+++ src/database/Wallets.kt
@@ -7,14 +7,14 @@ import org.jetbrains.exposed.dao.LongIdTable
 
 object Wallets : LongIdTable() {
     val currency = enumeration("currency", Currency::class)
-    val ballance = long("ballance")
+    val ballance = double("ballance")
     val customerId = long("customerId").references(Customers.id)
 }
 
 class Wallet(
     id: EntityID<Long>,
     val currency: Currency,
-    var ballance: Long,
+    var ballance: Double,
     val customerId: Long
 ) : LongEntity(id) {
     companion object : LongEntityClass<Wallet>(Wallets)
